Allow unknown local users
They will be considered to not be in any group and thus not be able
to issue for root. This allows the ud-ldap setup where users that are
not able to log into the machine directly by not exporting them to the
machine in question to work properly for issuance.